Updated Gantt.ts to support durations in seconds
Since the start time supports fractions of a day, it's also required that the duration supports fractions of a day. If not, there will in some cases be errors when plotting the graphs. Also, this change now opens up for using the Gantt chart for other things than just projects, e.g. showing ticket-system status or manufactoring flows